pyproject.toml
[project]
name = "topic-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.19.0",
    "aiogram-dialog>=2.3.1",
    "alembic>=1.15.2",
    "arxiv>=2.2.0",
    "dynaconf>=3.2.10",
    "fluentogram>=1.1.10",
    "gradio>=5.24.0",
    "nats-py>=2.10.0",
    "ormsgpack>=1.9.1",
    "psycopg>=3.2.6",
    "psycopg-pool>=3.2.6",
    "redis>=5.2.1",
    "scikit-learn>=1.6.1",
    "taskiq-nats>=0.5.1",
    "taskiq-redis>=1.0.4",
    "torch>=2.6.0",
    "transformers>=4.51.1",
]
